Fixed warnings for the base executable, without tests

+++ projects/miopen/src/include/miopen/kthvalue/problem_description.hpp
@@ -91,7 +91,7 @@ struct FwdProblemDescription : ProblemDescriptionBase
         int32_t posOut = 0;
         for(auto i = 0U; i < inputDesc.GetLengths().size(); i++)
         {
-            if(i == dim)
+            if(i == static_cast<size_t>(dim))
             {
                 if(!keepDim)
                     continue;
